// src/teamcity-client/models/step.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamCity REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2018.1 (current)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// May contain unused imports in some cases
// @ts-ignore
import type { Properties } from './properties';

/**
 * A build step: the smallest unit of a CI/CD routine within a build configuration or template, encapsulating a set of indivisible build instructions.
 * @export
 * @interface Step
 */
export interface Step {
  /**
   * The internal read-only ID of the object. This property is inherited from the base PropEntity class.
   * @type {string}
   * @memberof Step
   */
  id?: string;
  /**
   * The public object name. Some objects, like build steps, support public names, while others, like build triggers, do not. This property is inherited from the base PropEntity class.
   * @type {string}
   * @memberof Step
   */
  name?: string;
  /**
   * The object type. This property is inherited from the base PropEntity class.
   * @type {string}
   * @memberof Step
   */
  type?: string;
  /**
   * Returns **rue** if the object is disabled and inactive; otherwise, **false**. This property is inherited from the base PropEntity class.
   * @type {boolean}
   * @memberof Step
   */
  disabled?: boolean;
  /**
   * Returns **rue** if the object is inherited from another object; otherwise, **false**. This property is inherited from the base PropEntity class.
   * @type {boolean}
   * @memberof Step
   */
  inherited?: boolean;
  /**
   * Returns the shortened (without the server URL) link to the current object. This property is inherited from the base PropEntity class.
   * @type {string}
   * @memberof Step
   */
  href?: string;
  /**
   *
   * @type {Properties}
   * @memberof Step
   */
  properties?: Properties;
  /**
   * The short description of this build step, displayed in the list of configuration steps in TeamCity UI.
   * @type {string}
   * @memberof Step
   */
  shortDescription?: string;
}