// src/teamcity-client/models/steps.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TeamCity REST API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2018.1 (current)
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
// May contain unused imports in some cases
// @ts-ignore
import type { Step } from './step';

/**
 * Represents a list of Step entities.
 * @export
 * @interface Steps
 */
export interface Steps {
  /**
   * The number of build steps in this collection.
   * @type {number}
   * @memberof Steps
   */
  count?: number;
  /**
   * Build steps owned by this collection.
   * @type {Array<Step>}
   * @memberof Steps
   */
  step?: Array<Step>;
}